implement inProgressUpdates
