+++ cmd/traceroute.go
@@ -14,7 +14,7 @@ var tracerouteCmd = &cobra.Command{
 	GroupID: "Measurements",
 	Short:   "Use the native traceroute command",
 	Long: `traceroute tracks the route packets taken from an IP network on their way to a given host. It utilizes the IP protocol's time to live (TTL) field and attempts to elicit an ICMP TIME_EXCEEDED response from each gateway along the path to the host.
-	
+
 Examples:
   # Traceroute google.com from 2 probes in New York
   traceroute google.com from New York --limit 2
@@ -37,10 +37,11 @@ Examples:
 
 		// Make post struct
 		opts = model.PostMeasurement{
-			Type:      "traceroute",
-			Target:    ctx.Target,
-			Locations: createLocations(ctx.From),
-			Limit:     ctx.Limit,
+			Type:              "traceroute",
+			Target:            ctx.Target,
+			Locations:         createLocations(ctx.From),
+			Limit:             ctx.Limit,
+			InProgressUpdates: inProgressUpdates(ctx.CI),
 			Options: &model.MeasurementOptions{
 				Protocol: protocol,
 				Port:     port,
